requirements.txt
groq==0.11.0
httpx==0.27.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
streamlit==1.35.0